+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ForwardTriggerEvents component.
 - [UIFader] FadeInCoroutine/FadeOutCoroutine methods that return the Coroutine handle.
-- RoundToPrecision math util
+- RoundToPrecision math util.
+- Timer component.
 
 ### Changed
 - DetectGround sets IsOnGround to false on disable.